+++ Frends.HTTP.Request/Frends.HTTP.Request/Definitions/CertificateStoreLocation.cs
@@ -1,0 +1,16 @@
+namespace Frends.HTTP.Request.Definitions;
+
+/// <summary>
+/// Certificate store location.
+/// </summary>
+public enum CertificateStoreLocation
+{
+    /// <summary>
+    /// The X.509 certificate store assigned to the current user.
+    /// </summary>
+    CurrentUser,
+    /// <summary>
+    /// The X.509 certificate store assigned to the local machine.
+    /// </summary>
+    LocalMachine
+}