+++ Frends.HTTP.SendAndReceiveBytes/Frends.HTTP.SendAndReceiveBytes/Definitions/Options.cs
@@ -84,6 +84,15 @@ public class Options
     public string CertificateThumbprint { get; set; }
 
     /// <summary>
+    /// Applicable only when Certificate Source is "CertificateStore".
+    /// Store location for the certificate.
+    /// </summary>
+    /// <example>CurrentUser</example>
+    [UIHint(nameof(Authentication), "", Authentication.ClientCertificate)]
+    [DefaultValue(CertificateStoreLocation.CurrentUser)]
+    public CertificateStoreLocation CertificateStoreLocation { get; set; }
+
+    /// <summary>
     /// Should the entire certificate chain be loaded from the certificate store and included in the request. Only valid when using Certificate Store as the Certificate Source 
     /// </summary>
     /// <example>true</example>